Visualize the power analysis result

Visualize the power analysis result

Usage

visualizePowerResult(
  power_result,
  celltypes,
  x_axis = "nindiv",
  y_axis = "ncell",
  col_group = "method",
  geneid,
  snpid
)

Arguments

power_result

a data frame contains power analysis result in different parameter settings.

celltypes

a vector of cell types that will be selected for visualization.

x_axis

a character that specifies the x axis. Default is the number of individuals.

y_axis

a character that specifies the y axis. Default is the number of cells per individual.

col_group

a character that specifies the color groups. Default is the eQTL model.

geneid

a character object contains geneid to be part of the plot title.

snpid

a character object contains snpid to be part of the plot title.

Value

a ggplot object

Examples

NULL
#> NULL